Moving all column types to a sub-module
