+++ declaredata_fuse/src/declaredata_fuse/column/base.py
