+++ declaredata_fuse/src/declaredata_fuse/column/derived.py
@@ -1,25 +1,7 @@
-from enum import Enum
 from dataclasses import dataclass
 
+from declaredata_fuse.column.operations import BinaryOp
 from declaredata_fuse.proto import sds_pb2
-
-
-class BinaryOp(Enum):
-    ADD = 0
-    SUB = 1
-    MUL = 2
-    DIV = 3
-
-    def __str__(self) -> str:
-        match self:
-            case BinaryOp.ADD:
-                return "+"
-            case BinaryOp.SUB:
-                return "-"
-            case BinaryOp.MUL:
-                return "*"
-            case BinaryOp.DIV:
-                return "/"
 
 
 @dataclass
