+++ declaredata_fuse/src/declaredata_fuse/column/drop.py
@@ -1,0 +1,4 @@
+from declaredata_fuse.column.base import Column
+
+
+DropColumn = str | Column
