+++ declaredata_fuse/src/declaredata_fuse/sql/__init__.py
@@ -1,5 +1,5 @@
 from declaredata_fuse import functions
-from declaredata_fuse.column_abc import Column
+from declaredata_fuse.column.base import Column
 from declaredata_fuse.dataframe import DataFrame
 from declaredata_fuse.window import Window, WindowSpec
